requirements-dev.txt
python-dateutil
flask
flask-login
flask-sqlalchemy
flask-wtf
flask-migrate
